Fix logic bug in capacity calculation
The bug:

Imagine we have a flow that goes according to the convention set by
commit b3da8f3:

```
 _      _
|P|--->|S|
|_| 3  |_|

```

In this case, we will have, on the position of the parent in the matrix
of horizontalFlow the number 3. But if we have the simetric situation:

```
 _      _
|S|<---|P|
|_|  3 |_|

```

Then we have to use the position of the *son* (and put there *-3*).
This difference was not accounted for on our switch-case statements.

Changes to be committed:
	modified:   src/main.c