Fix error on test importing
On the previous commit b2c7ea0 I tried to change the name of some tests
to standardize names. That backfired badly, because some tests had name
collisions. So, this commit changes run_tests.sh to use *.in as input
for the test, instead of relying on a name scheme.