Fix allocation bug in double matrix allocation
The bug was in a mistake on a calloc call. The bug is similar to the
following one: instead of the correct

```
  (int *) res = (int *) calloc(n, sizeof(int));
```

I was doing

```
  (int *) res = (int *) calloc(n, sizeof(int*));
```

In this example, the mistake has no repercussions, because in general

`sizeof(int) == sizeof(int*)`

But in the case of double matrix, that is of type int(**)[2], that made
a real difference in the pointer arithmetic. Namely, in test t2.in, we
had `matrix[1][3] = matrix[2][1]`, and that generated a loop, and when
trying to go back to the BFS parent it was generating a infinite loop.

The fix was correcting the sizes in the calloc call.

Changes to be committed:
	modified:   src/main.c